Cargué el client para hacer el inscripcion

+++ backend/dao/inscripcion.go
@@ -1,9 +1,9 @@
 package dao
 
 type Inscripcion struct {
-	ID        int `gorm:"primary_key;AUTO_INCREMENT"`
-	UsuarioID int `gorm:"type:int;not null"`
-	CursoID   int `gorm:"type:int;not null"`
+	ID        int64 `gorm:"primary_key;AUTO_INCREMENT"`
+	UsuarioID int64 `gorm:"type:int;not null"`
+	CursoID   int64 `gorm:"type:int;not null"`
 }
 
 type Inscripciones []Inscripcion
